organise LDP features by ecosystem, move suites to submods
