+++ ssi-ldp/src/suites/mod.rs
@@ -1,0 +1,21 @@
+#[cfg(feature = "aleo")]
+mod aleo;
+#[cfg(feature = "eip")]
+mod eip;
+#[cfg(feature = "solana")]
+mod solana;
+#[cfg(feature = "tezos")]
+mod tezos;
+#[cfg(feature = "w3c")]
+mod w3c;
+
+#[cfg(feature = "aleo")]
+pub use aleo::*;
+#[cfg(feature = "eip")]
+pub use eip::*;
+#[cfg(feature = "solana")]
+pub use solana::*;
+#[cfg(feature = "tezos")]
+pub use tezos::*;
+#[cfg(feature = "w3c")]
+pub use w3c::*;
